+++ app/src/main/java/org/gnucash/android/db/MigrationHelper.java
@@ -24,7 +24,6 @@ import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Environment;
-import android.preference.PreferenceManager;
 import android.text.TextUtils;
 import android.util.Log;
 
@@ -45,6 +44,8 @@ import org.gnucash.android.model.PeriodType;
 import org.gnucash.android.model.Recurrence;
 import org.gnucash.android.model.ScheduledAction;
 import org.gnucash.android.model.Transaction;
+import org.gnucash.android.util.PreferencesHelper;
+import org.gnucash.android.util.TimestampHelper;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
@@ -62,6 +63,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.TimeZone;
 
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParser;
@@ -651,7 +653,7 @@ public class MigrationHelper {
             //================================ END TABLE MIGRATIONS ================================
 
             // String timestamp to be used for all new created entities in migration
-            String timestamp = (new Timestamp(System.currentTimeMillis())).toString();
+            String timestamp = TimestampHelper.getUtcStringFromTimestamp(TimestampHelper.getTimestampFromNow());
 
             //ScheduledActionDbAdapter scheduledActionDbAdapter = new ScheduledActionDbAdapter(db);
             //SplitsDbAdapter splitsDbAdapter = new SplitsDbAdapter(db);
@@ -702,7 +704,7 @@ public class MigrationHelper {
             while (cursor.moveToNext()){
                 contentValues.clear();
                 Timestamp timestampT = new Timestamp(cursor.getLong(cursor.getColumnIndexOrThrow(TransactionEntry.COLUMN_TIMESTAMP)));
-                contentValues.put(TransactionEntry.COLUMN_CREATED_AT, timestampT.toString());
+                contentValues.put(TransactionEntry.COLUMN_CREATED_AT, TimestampHelper.getUtcStringFromTimestamp(timestampT));
                 long transactionId = cursor.getLong(cursor.getColumnIndexOrThrow(TransactionEntry._ID));
                 db.update(TransactionEntry.TABLE_NAME, contentValues, TransactionEntry._ID + "=" + transactionId, null);
 
@@ -1103,11 +1105,9 @@ public class MigrationHelper {
 
                 boolean exportAll = Boolean.parseBoolean(tokens[2]);
                 if (exportAll){
-                    params.setExportStartTime(Timestamp.valueOf(Exporter.TIMESTAMP_ZERO));
+                    params.setExportStartTime(TimestampHelper.getTimestampFromEpochZero());
                 } else {
-                    String lastExportTimeStamp = PreferenceManager.getDefaultSharedPreferences(GnuCashApplication.getAppContext())
-                            .getString(Exporter.PREF_LAST_EXPORT_TIME, Exporter.TIMESTAMP_ZERO);
-                    Timestamp timestamp = Timestamp.valueOf(lastExportTimeStamp);
+                    Timestamp timestamp = PreferencesHelper.getLastExportTime();
                     params.setExportStartTime(timestamp);
                 }
 
@@ -1152,10 +1152,9 @@ public class MigrationHelper {
                 String tag = cursor.getString(cursor.getColumnIndexOrThrow(ScheduledActionEntry.COLUMN_TAG));
                 String[] tokens = tag.split(";");
                 try {
-                    Timestamp timestamp = Timestamp.valueOf(tokens[2]);
+                    Timestamp timestamp = TimestampHelper.getTimestampFromUtcString(tokens[2]);
                 } catch (IllegalArgumentException ex) {
-                    tokens[2] = PreferenceManager.getDefaultSharedPreferences(GnuCashApplication.getAppContext())
-                            .getString(Exporter.PREF_LAST_EXPORT_TIME, Exporter.TIMESTAMP_ZERO);
+                    tokens[2] = TimestampHelper.getUtcStringFromTimestamp(PreferencesHelper.getLastExportTime());
                 } finally {
                     tag = TextUtils.join(";", tokens);
                 }
@@ -1180,8 +1179,45 @@ public class MigrationHelper {
         return oldVersion;
     }
 
+    public static Timestamp subtractTimeZoneOffset(Timestamp timestamp, TimeZone timeZone) {
+        final long millisecondsToSubtract = Math.abs(timeZone.getOffset(timestamp.getTime()));
+        return new Timestamp(timestamp.getTime() - millisecondsToSubtract);
+    }
+
     /**
-     * Upgrades the database to version 12.
+     * Upgrade database to version 12
+     * <p>
+     *     Change last_export_time Android preference to current value - N
+     *     where N is the absolute timezone offset for current user time zone.
+     *     For details see #467.
+     * </p>
+     * @param db SQLite database
+     * @return 12 if upgrade was successful, 11 otherwise
+     */
+    static int upgradeDbToVersion12(SQLiteDatabase db){
+        Log.i(MigrationHelper.LOG_TAG, "Upgrading database to version 12");
+
+        int oldVersion = 11;
+
+        try {
+
+            final Timestamp currentLastExportTime = PreferencesHelper.getLastExportTime();
+
+            final Timestamp updatedLastExportTime = subtractTimeZoneOffset(
+                    currentLastExportTime, TimeZone.getDefault());
+            PreferencesHelper.setLastExportTime(updatedLastExportTime);
+
+            oldVersion = 12;
+
+        } catch (Exception ignored){
+            // Do nothing: here oldVersion = 11.
+        }
+
+        return oldVersion;
+    }
+
+    /**
+     * Upgrades the database to version 13.
      * <p>This migration makes the following changes to the database:
      * <ul>
      *     <li>Adds support for multiple database for different books and one extra database for storing book info</li>
@@ -1193,11 +1229,11 @@ public class MigrationHelper {
      * </ul>
      * </p>
      * @param db SQlite database to be upgraded
-     * @return New database version, 12 if migration succeeds, 11 otherwise
+     * @return New database version, 13 if migration succeeds, 11 otherwise
      */
-    static int upgradeDbToVersion12(SQLiteDatabase db){
-        Log.i(DatabaseHelper.LOG_TAG, "Upgrading database to version 9");
-        int oldVersion = 11;
+    static int upgradeDbToVersion13(SQLiteDatabase db){
+        Log.i(DatabaseHelper.LOG_TAG, "Upgrading database to version 13");
+        int oldVersion = 12;
 
         db.beginTransaction();
         try {
@@ -1394,7 +1430,7 @@ public class MigrationHelper {
             db.execSQL("DROP TABLE " + SplitEntry.TABLE_NAME + "_bak");
 
             db.setTransactionSuccessful();
-            oldVersion = 12;
+            oldVersion = 13;
         } finally {
             db.endTransaction();
         }