+++ app/src/main/java/org/gnucash/android/ui/util/widget/CalculatorEditText.java
@@ -51,12 +51,12 @@ import java.util.Locale;
 
 /**
  * A custom EditText which supports computations and uses a custom calculator keyboard.
- * <p>Afer the view is inflated, make sure to call {@link #bindListeners(KeyboardView)}
- * with the view from your layout where the calculator keyboard should be displayed:</p>
+ * <p>After the view is inflated, make sure to call {@link #bindListeners(KeyboardView)}
+ * with the view from your layout where the calculator keyboard should be displayed.</p>
  * @author Ngewi Fet <ngewif@gmail.com>
  */
 public class CalculatorEditText extends EditText {
-    CalculatorKeyboard mCalculatorKeyboard;
+    private CalculatorKeyboard mCalculatorKeyboard;
 
     private Commodity mCommodity = Commodity.DEFAULT_COMMODITY;
     private Context mContext;
@@ -125,7 +125,6 @@ public class CalculatorEditText extends EditText {
         mCalculatorKeyboard = calculatorKeyboard;
         mContext = calculatorKeyboard.getContext();
         setOnFocusChangeListener(new OnFocusChangeListener() {
-            // NOTE By setting the on focus listener, we can show the custom keyboard when the edit box gets focus, but also hide it when the edit box loses focus
             @Override
             public void onFocusChange(View v, boolean hasFocus) {
                 if (hasFocus) {
@@ -163,14 +162,11 @@ public class CalculatorEditText extends EditText {
             }
         });
 
-        // Although it looks redundant having both onClickListener and OnTouchListener, removing
-        // one of them makes the standard keyboard show up in addition to the calculator one.
+        // Although this handler doesn't make sense, if removed, the standard keyboard
+        // shows up in addition to the calculator one when the EditText gets a touch event.
         setOnTouchListener(new OnTouchListener() {
             @Override
             public boolean onTouch(View v, MotionEvent event) {
-                if (!mCalculatorKeyboard.isCustomKeyboardVisible())
-                    mCalculatorKeyboard.showCustomKeyboard(v);
-
                 // XXX: Use dispatchTouchEvent()?
                 onTouchEvent(event);
                 return false;
@@ -181,14 +177,14 @@ public class CalculatorEditText extends EditText {
     }
 
     /**
-     * Initializes listeners on the edittext
+     * Initializes listeners on the EditText
      */
     public void bindListeners(KeyboardView keyboardView){
         bindListeners(new CalculatorKeyboard(mContext, keyboardView, mCalculatorKeysLayout));
     }
 
     /**
-     * Returns the calculator keyboard instantiated by this edittext
+     * Returns the calculator keyboard instantiated by this EditText
      * @return CalculatorKeyboard
      */
     public CalculatorKeyboard getCalculatorKeyboard(){
@@ -255,8 +251,8 @@ public class CalculatorEditText extends EditText {
     }
 
     /**
-     * Evaluates the arithmetic expression in the editText and sets the text property
-     * @return Result of arithmetic evaluation which is same as text displayed in edittext
+     * Evaluates the arithmetic expression in the EditText and sets the text property
+     * @return Result of arithmetic evaluation which is same as text displayed in EditText
      */
     public String evaluate(){
         String amountString = getCleanString();
@@ -332,7 +328,7 @@ public class CalculatorEditText extends EditText {
     }
 
     /**
-     * Set the text to the value of {@code amount} formatted according to the locale
+     * Set the text to the value of {@code amount} formatted according to the locale.
      * <p>The number of decimal places are determined by the currency set to the view, and the
      * decimal separator is determined by the device locale. There are no thousandths separators.</p>
      * @param amount BigDecimal amount