+++ CHANGELOG.md
@@ -6,6 +6,16 @@ Version 2.1.0 *(2016-xx-xx)*
 * Improved: Scheduled transactions now have more accurate timestamps
 * Improved: Generate all scheduled transactions even if a scheduled is missed (e.g. device off)
 
+Version 2.0.6 *(2016-02-20)*
+----------------------------
+* Fixed: Saving transaction gets slower with increase in size of database
+* Fixed: Imbalance amount wrongly computed in split editor (for some accounts)
+* Fixed: Amount text boxes in split editor sometimes do not get focus
+* Fixed: Crash when saving account with no transfer account selected
+* Fixed: Crash when creating a new transaction with no transfer account
+* Fixed: All transactions are always exported for some time zones
+* Improved: Add translation for Japanese. Updated Italian and Russian
+
 Version 2.0.5 *(2015-12-12)*
 ----------------------------
 * Fixed: Wrong decimal formatting in multi-currency transactions